REGION 모델로 옮기기
