+++ src/main/kotlin/kr/mashup/bangwidae/asked/service/place/PlaceService.kt
@@ -2,6 +2,7 @@ package kr.mashup.bangwidae.asked.service.place
 
 import kr.mashup.bangwidae.asked.exception.DoriDoriException
 import kr.mashup.bangwidae.asked.exception.DoriDoriExceptionType
+import kr.mashup.bangwidae.asked.model.Region
 import org.springframework.stereotype.Service
 
 @Service
