+++ src/main/kotlin/kr/mashup/bangwidae/asked/service/place/ReverseGeocodeConverter.kt
@@ -1,14 +1,7 @@
 package kr.mashup.bangwidae.asked.service.place
 
+import kr.mashup.bangwidae.asked.model.Region
+
 interface ReverseGeocodeConverter {
     fun convert(longitude: Double, latitude: Double): Region
 }
-
-data class Region(
-    val 국가: Nationality?,
-    val 도: String?,
-    val 시: String?,
-    val 군구: String?,
-    val 읍면동: String?,
-    val 리: String?,
-)
